Add --path-prefix="prefix" option to ureadahead
to support limiting tracing to operations affecting a single directory.
It is useful for tracing the processes which are inside a container.

This CL also changes the behavior of 'ureadahead --force-trace
/path/to/mount/point'. Previously, it was exactly the same as
'ureadahead --force-trace' and sometimes wrote out multiple files
to /var/lib/ureadahead/. With this CL, it only creates/updates a
single file, /var/lib/ureadahead/path.to.mount.point.pack.

Link: https://chromium.googlesource.com/chromiumos/overlays/chromiumos-overlay/+/10dd5a6b06c79a3c80e63bb9c3caa60a01a4fb74

Tested-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@google.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>