ureadahead: Add pack-file and path-prefix-filter options
This change adds two command line options allowing multiple instances of
ureadahead to be more isolated:
* --pack-file: Overrides the path to the pack file instead of
determining it from the underlying device which an unrelated pack file
might already exist for. This overridden path gets used for tracing,
dumping the pack file as well as reading ahead.
* --path-prefix-filter: Prefix that the paths read from the debugfs
trace file have to start with for them to be retained in the pack file.

These new options will be used by the upstart job starting ureadahead
when the login screen appears in order to preload ARCVM files.

Link: https://chromium.googlesource.com/chromiumos/overlays/chromiumos-overlay/+/4df8fcf81914ef22cb4e716186d2bdb99a6d5944

Tested-by: Philippe Liard <pliard@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>